% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_data_frame}
\alias{check_data_frame}
\title{Check that an object is a data frame}
\usage{
check_data_frame(x, name)
}
\arguments{
\item{x}{Object to be checked.}

\item{name}{Character scalar. Name of the object, used in error messages.}
}
\value{
Invisibly returns \code{TRUE} if the check passes.
}
\description{
Verifies that an object is a data frame.
}
