% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_data_frame_path}
\alias{check_data_frame_path}
\title{Check that a directory exists}
\usage{
check_data_frame_path(path)
}
\arguments{
\item{path}{Character scalar. Path to a directory.}
}
\value{
Invisibly returns \code{TRUE} if the check passes.
}
\description{
Verifies that a directory exists at the specified path.
}
