% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_logical_value}
\alias{check_logical_value}
\title{Check that an object is a logical scalar}
\usage{
check_logical_value(x, name)
}
\arguments{
\item{x}{Object to be checked.}

\item{name}{Character scalar. Name of the object, used in error messages.}

\item{allow_na}{Logical. Should missing values be allowed? Defaults to
\code{FALSE}.}
}
\value{
Invisibly returns \code{TRUE} if the check passes.
}
\description{
Verifies that an object is a logical vector of length one.
}
