% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_numeric_value}
\alias{check_numeric_value}
\title{Check that an object is a numeric scalar}
\usage{
check_numeric_value(x, name, allow_na = FALSE)
}
\arguments{
\item{x}{Object to be checked.}

\item{name}{Character scalar. Name of the object, used in error messages.}

\item{allow_na}{Logical. Should missing values be allowed? Defaults to
\code{FALSE}.}
}
\value{
Invisibly returns \code{TRUE} if the check passes.
}
\description{
Verifies that an object is a numeric vector of length one and optionally
allows missing values.
}
