% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_same_length}
\alias{check_same_length}
\title{Check that two objects have the same length}
\usage{
check_same_length(x, y, name_x, name_y)
}
\arguments{
\item{x}{First object.}

\item{y}{Second object.}

\item{name_x}{Character scalar. Name of the first object.}

\item{name_y}{Character scalar. Name of the second object.}
}
\value{
Invisibly returns \code{TRUE} if the check passes.
}
\description{
Verifies that two vectors have the same length.
}
